back-end/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
cryptography==40.0.2
defusedxml==0.7.1
dj-rest-auth==4.0.0
Django==3.2.13
django-allauth==0.52.0
django-cors-headers==3.14.0
djangorestframework==3.14.0
duckdb==0.8.0
filelock==3.12.0
frozenlist==1.3.3
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.6
pandas==2.0.1
Pillow==9.5.0
pyarrow==12.0.0
pycparser==2.21
PyJWT==2.7.0
pylance==0.4.13
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2022.6
requests==2.30.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
sympy==1.12
torch==2.0.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
yarl==1.9.2
